# 스마트검색 제목_작성자_가격적용

+++ src/main/java/store/itcen/community/postapi/dto/SearchDTO.java
@@ -9,7 +9,7 @@ import lombok.*;
 public class SearchDTO {
     private String searchTitle;
     private String searchWriter;
-    private int searchPriceMin;
-    private int searchPriceMax;
+    private Long searchPriceMin;
+    private Long searchPriceMax;
 
 }
